+++ rog-dbus/src/lib.rs
@@ -1,5 +1,4 @@
 pub static DBUS_NAME: &str = "org.asuslinux.Daemon";
-pub static DBUS_NAME_GFX: &str = "org.supergfxctl.Daemon";
 pub static DBUS_PATH: &str = "/org/asuslinux/Daemon";
 pub static DBUS_IFACE: &str = "org.asuslinux.Daemon";